Move examples to their own module, to reduce clutter
